Fixed wrong airline routes displayed #72
Fixed airports that shouldn't popup in the Airport.php  query that was targeting the wrong way.

Then also fixed the collection->addFlights() function to create two seperate relations. As making 2 level deep relationship messed badly with the end result.

Hence the view now checks for airlines and then flights with a condition it's that specific airline. Though since all queries are done, it's still an optimized way to fetch and filter.