feat: record aircraft types per flight
In this update I've added the backend groundwork for fetching flight aircraft and putting them in a separate table. Also changed so today's frontend uses the new dataset.

The aircraft_icao field remains in the flight table for duplication purposes due to upsert not being able to return ids to do this in one go.